Se agregaron metodos crud de servicio consultas.

+++ servicioConsultas/src/main/java/com/iamf/servicioConsultas/repositories/TipoServicioRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoServicioRepo extends JpaRepository<TipoServicio, Integer> {
+public interface TipoServicioRepo extends JpaRepository<TipoServicio, Long> {
 }
